+++ hmda/src/main/resources/persistence.conf
@@ -19,10 +19,10 @@ akka {
     }
     serialization-bindings {
       "hmda.model.institution.Institution" = institution-commands
-      "hmda.persistence.institution.InstitutionPersistence$CreateInstitution" = institution-commands
-      "hmda.persistence.institution.InstitutionPersistence$ModifyInstitution" = institution-commands
-      "hmda.persistence.institution.InstitutionPersistence$GetInstitution" = institution-commands
-      "hmda.persistence.institution.InstitutionPersistence$DeleteInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$CreateInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$ModifyInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$GetInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$DeleteInstitution" = institution-commands
 
       "hmda.messages.institution.InstitutionEvents$InstitutionCreated" = institution-events
       "hmda.messages.institution.InstitutionEvents$InstitutionModified" = institution-events
